+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [11.0.1] - UNRELEASED
 
 - Added `ApplicationUtility.GetDisplayName(string)` as runtime version for `EngineUtility.GetDisplayName(string)`.
+- Changed `Linting Settings` to come with the default rules set.
 - Fixed some documentation typos and missing links.
 - Fixed edge case on `PlayerLoopSystem` that caused already relevant events to not be registered correctly.